feat: all layout components supports typescript
